#758: refactor 重构 LLM 调用为双层结构
- 新建 lib/chat/base-llm.js：统一 HTTP 调用层（~360行）
- 重构 lib/llm-client.js：删除 HTTP 逻辑，委托 BaseLLM（1124→701行）
- 简化 lib/internal-llm-service.js：直接调用 BaseLLM（496→433行）
- 改造 lib/simple-llm-client.js：改为 BaseLLM 的 re-export（357→18行）
- 修复 InternalLLMService.call() 未透传 thinking/reasoning/tools 参数
