+++ lib/internal-llm-service.js
@@ -7,12 +7,11 @@
  * - 通过 expertId 获取专家关联的模型配置（优先使用反思模型）
  * - 使用较低温度（默认 0.3），输出更确定
  * - 支持 JSON 输出格式和 Schema 校验
- * - 底层使用 LLMClient 进行 API 调用
+ * - 底层使用 BaseLLM 进行 API 调用
  */
 
 import logger from './logger.js';
-import { retryWithBackoff, isRetryableError } from './llm-retry.js';
-import LLMClient from './llm-client.js';
+import { callWithRetry as baseCallWithRetry } from './chat/base-llm.js';
 import ConfigLoader from './config-loader.js';
 
 class InternalLLMService {
@@ -29,36 +28,12 @@ class InternalLLMService {
     this.maxRetries = options.maxRetries ?? 3;
     this.timeout = options.timeout ?? 90000;
 
-    // 缓存层 1: 模型配置缓存（减少数据库查询）
     this.modelConfigCache = new Map();
-
-    // 缓存层 2: HTTP 客户端缓存（复用 TCP 连接，减少网络开销）
-    this.httpClientCache = new Map();
-
-    // 共享 ConfigLoader 实例
     this.configLoader = new ConfigLoader();
   }
 
   /**
-   * 获取指定专家的 HTTP 客户端实例
-   * @param {string} expertId - 专家ID
-   * @returns {Promise<LLMClient>}
-   */
-  async getHttpClient(expertId) {
-    if (this.httpClientCache.has(expertId)) {
-      return this.httpClientCache.get(expertId);
-    }
-    const httpClient = new LLMClient(this.configLoader, expertId);
-    await httpClient.loadConfig();
-    this.httpClientCache.set(expertId, httpClient);
-    return httpClient;
-  }
-
-  /**
    * 估算文本的 token 数
-   * 中文字符 ≈ 1.5 token，其他字符 ≈ 0.25 token
-   * @param {string} text
-   * @returns {number}
    */
   estimateTokens(text) {
     if (!text) return 0;
@@ -69,10 +44,6 @@ class InternalLLMService {
 
   /**
    * 根据模型上下文限制截断文本
-   * @param {Object} model - 模型配置
-   * @param {string} systemPrompt
-   * @param {string} userPrompt
-   * @returns {{ systemPrompt: string, userPrompt: string, truncated: boolean }}
    */
   truncateForContext(model, systemPrompt, userPrompt) {
     const maxTokens = model.max_tokens || 128000;
@@ -83,7 +54,6 @@ class InternalLLMService {
     const systemTokens = this.estimateTokens(systemPrompt);
     const userTokens = this.estimateTokens(userPrompt);
     const totalTokens = systemTokens + userTokens;
-
     const limit = Math.floor(maxInput * safetyRatio);
 
     if (totalTokens <= limit) {
@@ -92,15 +62,12 @@ class InternalLLMService {
 
     const availableForUser = limit - systemTokens;
     if (availableForUser <= 0) {
-      logger.warn(`[InternalLLMService] System prompt alone exceeds context limit (${systemTokens} > ${limit})`);
       return { systemPrompt, userPrompt, truncated: true };
     }
 
     const estimatedCharsPerToken = userTokens > 0 ? userPrompt.length / userTokens : 4;
     const safeChars = Math.floor(availableForUser * estimatedCharsPerToken * 0.9);
-
-    const truncated = userPrompt.substring(0, safeChars) + '\n\n[... 文本因超出模型上下文限制已截断 ...]';
-    logger.warn(`[InternalLLMService] Input truncated: ${totalTokens} tokens → ~${limit} tokens (model max=${maxTokens}, output reserve=${maxOutput})`);
+    const truncated = userPrompt.substring(0, safeChars) + '\n\n[... 文本超出模型上下文限制已截断 ...]';
 
     return { systemPrompt, userPrompt: truncated, truncated: true };
   }
@@ -334,16 +301,11 @@ const response = await this.callWithRetry(model, messages, {
     return response.content;
 }
 
-  /**
-   * 通用 LLM 调用方法 - 使用 LLMClient
-   * @param {Object} model - 模型配置
-   * @param {Array} messages - 消息数组
-   * @param {Object} options - 可选参数
-   * @param {string} options.expertId - 专家ID（用于获取 LLMClient）
-   * @returns {Promise<Object>} 包含 content 和 usage 的响应
+    /**
+   * 通用 LLM 调用方法 - 使用 BaseLLM
    */
   async call(model, messages, options = {}) {
-    const { expertId, enableThinking } = options;
+    const { enableThinking } = options;
     const shouldDisableThinking = enableThinking === false || enableThinking === undefined;
 
     const finalMessages = shouldDisableThinking && messages.length > 0
@@ -361,17 +323,17 @@ const response = await this.callWithRetry(model, messages, {
       messages_count: messages.length,
     });
 
-    // 获取对应的 HTTP 客户端实例
-    const httpClient = expertId ? await this.getHttpClient(expertId) : await this.getAnyHttpClient();
-
-    // 使用 LLMClient.call() 进行非流式调用
-    const result = await httpClient.call(model, finalMessages, {
+    const result = await baseCallWithRetry(model, finalMessages, {
       temperature: options.temperature ?? this.defaultTemperature,
       top_p: options.top_p ?? 1.0,
       frequency_penalty: options.frequency_penalty ?? 0.0,
       presence_penalty: options.presence_penalty ?? 0.0,
       max_tokens: options.max_tokens || model.max_output_tokens || 8192,
       response_format: options.response_format,
+      thinking: options.thinking,
+      reasoning: options.reasoning,
+      tools: options.tools,
+      tool_choice: options.tool_choice,
       timeout: options.timeout || this.timeout,
     });
 
@@ -383,32 +345,10 @@ const response = await this.callWithRetry(model, messages, {
   }
 
   /**
-   * 获取任意一个缓存的 HTTP 客户端实例
-   * @returns {Promise<LLMClient>}
-   */
-  async getAnyHttpClient() {
-    const first = this.httpClientCache.values().next().value;
-    if (first) return first;
-    throw new Error('No HTTP client available. Call getHttpClient(expertId) first.');
-  }
-
-  /**
-   * 带重试机制的 LLM 调用（使用公共重试模块）
-   * @param {Object} model - 模型配置
-   * @param {Array} messages - 消息数组
-   * @param {Object} options - 可选参数
-   * @returns {Promise<Object>} LLM 响应
+   * 带重试机制的 LLM 调用（BaseLLM 已内置重试，这里保留兼容）
    */
   async callWithRetry(model, messages, options = {}) {
-    return retryWithBackoff(
-      () => this.call(model, messages, options),
-      {
-        maxRetries: this.maxRetries,
-        baseDelayMs: 2000,
-        maxDelayMs: 30000,
-        loggerPrefix: '[InternalLLMService]',
-      }
-    );
+    return this.call(model, messages, options);
   }
 
   /**
@@ -482,15 +422,13 @@ const response = await this.callWithRetry(model, messages, {
    * 清除模型缓存
    * @param {string} key - 缓存键（可选，不传则清除所有）
    */
-  clearCache(key = null) {
+    clearCache(key = null) {
     if (key) {
       this.modelConfigCache.delete(key);
-      this.httpClientCache.delete(key);
     } else {
       this.modelConfigCache.clear();
-      this.httpClientCache.clear();
     }
- }
+  }
 }
 
 export default InternalLLMService;
