+++ lib/llm-client.js
@@ -1,13 +1,11 @@
 /**
- * LLM Client
+ * LLM Client (Expert Layer)
  * 支持二分心智架构（Expressive Mind + Reflective Mind）
- * 两种心智使用不同的模型配置
+ * 底层调用 BaseLLM 进行 HTTP 通信
  */
 
-import https from 'https';
-import http from 'http';
 import logger from './logger.js';
-import { retryWithBackoff, isRetryableError } from './llm-retry.js';
+import { call, callWithRetry, callStream } from './chat/base-llm.js';
 
 class LLMClient {
   /**
@@ -195,153 +193,16 @@ class LLMClient {
    * @returns {Promise<Object>} 包含 content、reasoning_content 和 usage 的响应
    */
   async call(model, messages, options = {}) {
-    const userId = options.user_id || 'anonymous';
-    const requestId = this._generateRequestId(userId);
-    
-    // 处理多模态消息格式（传入模型配置以判断是否支持多模态）
     const processedMessages = this.processMultimodalMessages(messages, model);
 
-    const requestBody = JSON.stringify({
-      model: model.model_name,
-      messages: processedMessages,
-      temperature: options.temperature ?? 0.7,
-      top_p: options.top_p ?? 1.0,
-      frequency_penalty: options.frequency_penalty ?? 0.0,
-      presence_penalty: options.presence_penalty ?? 0.0,
-      max_tokens: options.max_output_tokens || model.max_output_tokens || 32768,
-      ...(options.tools && options.tools.length > 0 && { tools: options.tools }),
-      ...(options.tools && options.tools.length > 0 && options.tool_choice && { tool_choice: options.tool_choice }),
-      ...(options.response_format && { response_format: options.response_format }),
-      // DeepSeek 思考模式
-      ...(options.thinking && { thinking: options.thinking }),
-      // OpenAI 推理模式
-      ...(options.reasoning && { reasoning: options.reasoning }),
-    });
-
-    const url = new URL(model.base_url);
-    const isHttps = url.protocol === 'https:';
-    const httpModule = isHttps ? https : http;
-
-    // 优先使用 options 中的 timeout（毫秒），否则使用模型配置的 timeout（秒）转换为毫秒，默认 120 秒
-    const timeoutSec = model.timeout || 120;
-    const timeoutValue = options.timeout || (timeoutSec * 1000);
-
-    // 使用 provider 配置的 user_agent，如果未配置则使用默认值
-    const userAgent = model.user_agent || 'Version: 5.10.0 (c3d4709c)';
-    
-    const requestOptions = {
-      hostname: url.hostname,
-      port: url.port || (isHttps ? 443 : 80),
-      path: `${url.pathname}/chat/completions`,
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${model.api_key}`,
-        'Content-Length': Buffer.byteLength(requestBody),
-        'Connection': 'keep-alive',
-        'User-Agent': userAgent,
-      },
-      timeout: timeoutValue,
-      agent: new (isHttps ? https : http).Agent({
-        keepAlive: true,
-        keepAliveMsecs: 30000,
-        timeout: timeoutValue,
-      }),
-    };
-
-    // 详细日志：区分反思请求和正常请求
     const callType = options._reflective ? '反思' : '非流式';
     logger.info(`[LLMClient] 开始${callType}调用:`, {
       model_name: model.model_name,
-      url: `${requestOptions.hostname}:${requestOptions.port}${requestOptions.path}`,
-      timeout: timeoutValue,
-      timeout_source: options.timeout ? 'options' : (model.timeout ? 'model' : 'default'),
-      body_length: requestBody.length,
       messages_count: messages.length,
       temperature: options.temperature ?? 0.7,
     });
 
-return new Promise((resolve, reject) => {
-      const startTime = Date.now();
-      const req = httpModule.request(requestOptions, (res) => {
-        let data = '';
-        
-        res.on('data', (chunk) => {
-          data += chunk;
-        });
-        
-        res.on('end', () => {
-          // 请求完成后注销
-          this._unregisterRequest(requestId);
-          
-          try {
-            if (res.statusCode !== 200) {
-              logger.error('[LLMClient] 非流式调用失败:', {
-                status_code: res.statusCode,
-                response: data.substring(0, 500),
-                model: model.model_name,
-              });
-              reject(new Error(`HTTP ${res.statusCode}: ${data}`));
-              return;
-            }
-            
-            const response = JSON.parse(data);
-            const content = response.choices?.[0]?.message?.content;
-            const toolCalls = response.choices?.[0]?.message?.tool_calls;
-            const reasoningContent = response.choices?.[0]?.message?.reasoning_content;
-            
-            const duration = Date.now() - startTime;
-            logger.debug(`LLM call completed in ${duration}ms`, {
-              model: model.model_name,
-              tokens: response.usage?.total_tokens,
-              has_reasoning: !!reasoningContent,
-            });
-            
-            resolve({
-              content,
-              toolCalls,
-              reasoningContent, // DeepSeek 思考内容
-              usage: response.usage,
-              model: model.model_name,
-            });
-          } catch (error) {
-            logger.error('[LLMClient] 解析响应失败:', {
-              error: error.message,
-              response_preview: data.substring(0, 200),
-            });
-            reject(new Error(`Parse error: ${error.message}`));
-          }
-        });
-      });
-
-      req.on('error', (error) => {
-        // 错误时也注销
-        this._unregisterRequest(requestId);
-        logger.error('[LLMClient] 请求错误:', {
-          error: error.message,
-          error_code: error.code,
-          model: model.model_name,
-          url: `${requestOptions.hostname}:${requestOptions.port}${requestOptions.path}`,
-        });
-        reject(error);
-      });
-
-      req.on('timeout', () => {
-        this._unregisterRequest(requestId);
-        logger.error('[LLMClient] 请求超时:', {
-          timeout: requestOptions.timeout,
-          model: model.model_name,
-        });
-        req.destroy();
-        reject(new Error('Request timeout'));
-      });
-      
-      // 注册活跃请求，用于中止
-      this._registerRequest(requestId, req);
-      
-      req.write(requestBody);
-      req.end();
-    });
+    return call(model, processedMessages, options);
   }
 
   /**
@@ -352,15 +213,8 @@ return new Promise((resolve, reject) => {
    * @returns {Promise<Object>} LLM 响应
    */
   async callWithRetry(model, messages, options = {}) {
-    return retryWithBackoff(
-      () => this.call(model, messages, options),
-      {
-        maxRetries: 3,
-        baseDelayMs: 10000,   // LLMClient 使用较长的延迟，避免对本地 Ollama 等服务造成压力
-        maxDelayMs: 120000,
-        loggerPrefix: '[LLMClient]',
-      }
-    );
+    const processedMessages = this.processMultimodalMessages(messages, model);
+    return callWithRetry(model, processedMessages, options);
   }
 
   /**
@@ -695,295 +549,18 @@ return new Promise((resolve, reject) => {
    * @returns {Promise<void>}
    */
   async callStream(model, messages, options = {}) {
-    const userId = options.user_id || 'anonymous';
-    const requestId = this._generateRequestId(userId);
-    
-    const { tools, onDelta, onReasoningDelta, onToolCall, onUsage } = options;
-
-    // 处理多模态消息格式（传入模型配置以判断是否支持多模态）
     const processedMessages = this.processMultimodalMessages(messages, model);
-
-    // 分析消息构成
-    const messageAnalysis = this.analyzeMessages(processedMessages);
-    
-    // 估算当前对话的 token 数量
-    const estimatedTokens = this.estimateMessagesTokens(messages);
-    const maxContext = model.max_tokens || 128000;
-    const ratio = (estimatedTokens / maxContext * 100).toFixed(2);
+    const { tools } = options;
 
     logger.info('[LLMClient] 开始流式调用:', {
       model_name: model.model_name,
-      base_url: model.base_url,
-      has_api_key: !!model.api_key,
-      api_key_prefix: model.api_key?.substring(0, 10) + '...',
       messages_count: processedMessages.length,
       has_tools: !!(tools && tools.length > 0),
       has_thinking: !!options.thinking,
       has_reasoning: !!options.reasoning,
-      estimated_tokens: estimatedTokens,
-      max_context: maxContext,
-      context_ratio: ratio + '%',
-      message_analysis: messageAnalysis,
     });
 
-    const requestBody = JSON.stringify({
-      model: model.model_name,
-      messages: processedMessages,
-      temperature: options.temperature ?? 0.7,
-      top_p: options.top_p ?? 1.0,
-      frequency_penalty: options.frequency_penalty ?? 0.0,
-      presence_penalty: options.presence_penalty ?? 0.0,
-      max_tokens: options.max_output_tokens || model.max_output_tokens || 32768,
-      stream: true,
-      stream_options: { include_usage: true },  // 请求流式响应中包含 usage 信息
-      ...(tools && tools.length > 0 && { tools }),
-      ...(tools && tools.length > 0 && options.tool_choice && { tool_choice: options.tool_choice }),
-      // DeepSeek 思考模式
-      ...(options.thinking && { thinking: options.thinking }),
-      // OpenAI 推理模式
-      ...(options.reasoning && { reasoning: options.reasoning }),
-    });
-
-    const url = new URL(model.base_url);
-    const isHttps = url.protocol === 'https:';
-    const httpModule = isHttps ? https : http;
-
-    // 使用数据库中配置的 timeout（秒）转换为毫秒，默认 120 秒
-    const timeoutSec = model.timeout || 120;
-    const streamTimeout = timeoutSec * 1000;
-    
-    // 使用 provider 配置的 user_agent，如果未配置则使用默认值
-    const userAgent = model.user_agent || 'Version: 5.10.0 (c3d4709c)';
-    
-    const requestOptions = {
-      hostname: url.hostname,
-      port: url.port || (isHttps ? 443 : 80),
-      path: `${url.pathname}/chat/completions`,
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${model.api_key}`,
-        'Content-Length': Buffer.byteLength(requestBody),
-        'Accept': 'text/event-stream',
-        'Connection': 'keep-alive',
-        'User-Agent': userAgent,
-      },
-      timeout: streamTimeout,
-      // 添加 agent 配置以支持 keep-alive
-      agent: new (isHttps ? https : http).Agent({
-        keepAlive: true,
-        keepAliveMsecs: 30000,
-        timeout: streamTimeout,
-      }),
-    };
-    
-    logger.debug('[LLMClient] 请求详情:', {
-      url: `${requestOptions.hostname}:${requestOptions.port}${requestOptions.path}`,
-      timeout: requestOptions.timeout,
-      timeout_sec: Math.round(requestOptions.timeout / 1000) + 's',
-      body_length: requestBody.length,
-    });
-
-    return new Promise((resolve, reject) => {
-      const req = httpModule.request(requestOptions, (res) => {
-        if (res.statusCode !== 200) {
-          let errorData = '';
-          res.on('data', chunk => errorData += chunk);
-          res.on('end', () => {
-            reject(new Error(`HTTP ${res.statusCode}: ${errorData}`));
-          });
-          return;
-        }
-
-        let buffer = '';
-        let accumulatedToolCalls = {};  // 累积工具调用（按 index 累积）
-        let toolCallsSent = false;  // 标记工具调用是否已发送，防止重复
-        let pendingContent = '';  // 当前累积的文本（用于捕获工具调用的 context）
-        let toolCallContexts = {};  // 按工具 index 存储 context
-        
-        res.on('data', (chunk) => {
-          buffer += chunk.toString();
-          
-          // 处理 SSE 数据行
-          const lines = buffer.split('\n');
-          buffer = lines.pop(); // 保留不完整的最后一行
-          
-          for (const line of lines) {
-            if (line.startsWith('data: ')) {
-              const data = line.slice(6).trim();
-              
-              if (data === '[DONE]') {
-                // 流结束，处理累积的工具调用（如果尚未发送）
-                if (!toolCallsSent) {
-                  const finalToolCalls = Object.values(accumulatedToolCalls)
-                    .sort((a, b) => a.index - b.index)
-                    .map((tc, i) => ({
-                      // 如果 LLM 没有返回 id，生成一个唯一的 tool_call_id
-                      id: tc.id || `tool_call_${Date.now()}_${i}`,
-                      type: tc.type || 'function',
-                      function: {
-                        name: tc.function?.name || '',
-                        arguments: tc.function?.arguments || '',
-                      },
-                      context: toolCallContexts[tc.index] || '',  // 附加 context
-                    }));
-                  
-                  if (finalToolCalls.length > 0) {
-                    toolCallsSent = true;
-                    logger.info('[LLMClient] 流式工具调用完成:', {
-                      count: finalToolCalls.length,
-                      tools: finalToolCalls.map(tc => tc.function.name),
-                      contexts: finalToolCalls.map(tc => tc.context?.substring(0, 30) || ''),
-                    });
-                    onToolCall?.(finalToolCalls);
-                  }
-                }
-                resolve();
-                return;
-              }
-              
-              try {
-                const parsed = JSON.parse(data);
-
-                // 检测错误响应
-                if (parsed.error) {
-                  logger.error('[LLMClient] API 错误:', parsed.error);
-                  reject(new Error(parsed.error.message || JSON.stringify(parsed.error)));
-                  return;
-                }
-
-                const delta = parsed.choices?.[0]?.delta;
-
-                // 处理思考内容（DeepSeek reasoning_content）
-                if (delta?.reasoning_content) {
-                  onReasoningDelta?.(delta.reasoning_content);
-                }
-
-                if (delta?.content) {
-                  pendingContent += delta.content;  // 累积文本用于捕获 context
-                  onDelta?.(delta.content);
-                }
-
-                // 如果有 refusal，说明模型拒绝回答
-                if (delta?.refusal) {
-                  logger.warn('[LLMClient] 模型拒绝回答:', delta.refusal);
-                }
-                
-                // 累积工具调用（流式模式下是增量返回的）
-                if (delta?.tool_calls) {
-                  for (const tc of delta.tool_calls) {
-                    const idx = tc.index;
-                    if (!accumulatedToolCalls[idx]) {
-                      // ★ 关键：新的工具调用开始，捕获当前累积的文本作为 context
-                      accumulatedToolCalls[idx] = { index: idx, function: {} };
-                      toolCallContexts[idx] = pendingContent.trim();
-                      pendingContent = '';  // 清空，准备下一个工具的状态文本
-                      logger.debug(`[LLMClient] 捕获工具 #${idx} 的 context:`, toolCallContexts[idx]?.substring(0, 50));
-                    }
-                    // 累积各个字段
-                    if (tc.id) accumulatedToolCalls[idx].id = tc.id;
-                    if (tc.type) accumulatedToolCalls[idx].type = tc.type;
-                    if (tc.function?.name) {
-                      accumulatedToolCalls[idx].function.name = tc.function.name;
-                    }
-                    if (tc.function?.arguments) {
-                      accumulatedToolCalls[idx].function.arguments =
-                        (accumulatedToolCalls[idx].function.arguments || '') + tc.function.arguments;
-                    }
-                  }
-                }
-                
-                // 处理 usage 信息（流式响应的最后一个 chunk 会包含 usage）
-                if (parsed.usage) {
-                  onUsage?.(parsed.usage);
-                }
-              } catch (error) {
-                logger.warn('[LLMClient] 解析流式数据失败:', error.message);
-              }
-            }
-          }
-        });
-
-        res.on('end', () => {
-          // 请求完成后注销
-          this._unregisterRequest(requestId);
-          
-          // 确保在 res.on('end') 时也处理累积的工具调用（以防 [DONE] 没有被触发）
-          // 使用 toolCallsSent 标记防止重复发送
-          if (!toolCallsSent) {
-            const finalToolCalls = Object.values(accumulatedToolCalls)
-              .filter(tc => tc.function?.name)  // 确保有工具名称
-              .sort((a, b) => a.index - b.index)
-              .map((tc, i) => ({
-                // 如果 LLM 没有返回 id，生成一个唯一的 tool_call_id
-                id: tc.id || `tool_call_${Date.now()}_${i}`,
-                type: tc.type || 'function',
-                function: {
-                  name: tc.function?.name || '',
-                  arguments: tc.function?.arguments || '',
-                },
-                context: toolCallContexts[tc.index] || '',  // 附加 context
-              }));
-            
-            if (finalToolCalls.length > 0) {
-              toolCallsSent = true;
-              logger.info('[LLMClient] 流式结束(res.on end)，处理累积工具调用:', {
-                count: finalToolCalls.length,
-                tools: finalToolCalls.map(tc => tc.function.name),
-                contexts: finalToolCalls.map(tc => tc.context?.substring(0, 30) || ''),
-              });
-              onToolCall?.(finalToolCalls);
-            }
-          } else {
-            logger.debug('[LLMClient] res.on(end: 工具调用已发送，跳过');
-          }
-          resolve();
-        });
-
-        res.on('error', (error) => {
-          reject(error);
-        });
-      });
-
-      req.on('error', (error) => {
-        logger.error('[LLMClient] 流式请求错误:', {
-          error: error.message,
-          error_code: error.code,
-          model: model.model_name,
-          url: `${requestOptions.hostname}:${requestOptions.port}${requestOptions.path}`,
-        });
-        this._unregisterRequest(requestId);
-        reject(error);
-      });
-
-      req.on('timeout', () => {
-        this._unregisterRequest(requestId);
-        logger.error('[LLMClient] 流式请求超时:', {
-          timeout: requestOptions.timeout,
-          model: model.model_name,
-        });
-        req.destroy();
-        reject(new Error('Request timeout'));
-      });
-
-      // 监听 socket 级别错误
-      req.on('socket', (socket) => {
-        socket.on('error', (error) => {
-          this._unregisterRequest(requestId);
-          logger.error('[LLMClient] Socket 错误:', {
-            error: error.message,
-            error_code: error.code,
-            model: model.model_name,
-          });
-        });
-      });
-      
-      // 注册活跃请求，用于中止
-      this._registerRequest(requestId, req);
-      
-      req.write(requestBody);
-      req.end();
-    });
+    return callStream(model, processedMessages, options);
   }
 
   /**
